requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
beautifulsoup4==4.12.3
bentoml==1.3.7
bleach==6.1.0
Brotli==1.1.0
cattrs==23.1.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.0.0
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
deepmerge==2.0
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dukpy==0.4.0
dulwich==0.22.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fqdn==1.5.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.6.1
future==1.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-ws==0.6.2
huggingface-hub==0.24.6
idna==3.7
importlib-metadata==6.11.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.26.0
iso8601==2.1.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.8
multidict==6.1.0
mutagen==1.47.0
mutf8==1.0.6
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
nvidia-ml-py==11.525.150
openai==1.47.0
openllm==0.6.11
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pip-requirements-parser==32.0.1
platformdirs==4.2.2
pooch==1.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.36
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.9.0
PyAudio==0.2.14
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pynvim==0.5.0
pyparsing==3.1.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
PyYAML==6.0.1
pyzmq==26.0.3
questionary==2.0.1
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.19.1
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
Send2Trash==1.8.3
setuptools==71.1.0
shellingham==1.5.4
simple-di==0.1.5
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.39.2
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.0
tomli_w==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.2
uv==0.4.18
uvicorn==0.31.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
you-get==0.4.1730
yt-dlp==2024.10.22
zipp==3.20.2
